Fix long to Int problem

+++ src/main/java/org/example/rmi/ChatService.java
@@ -11,8 +11,8 @@ import java.util.List;
 public interface ChatService extends Remote {
     void createChat(Chat chat) throws RemoteException;
     List<Chat> getAllChats() throws RemoteException;
-    void subscribeUserToChat(Long userId, Long chatId) throws RemoteException;
-    void unsubscribeUserFromChat(Long userId, Long chatId) throws RemoteException;
+    void subscribeUserToChat(int userId, Long chatId) throws RemoteException;
+    void unsubscribeUserFromChat(int userId, Long chatId) throws RemoteException;
     void sendMessageToChat(Long chatId, String message) throws RemoteException;
     void sendMessage(String message, User sender) throws RemoteException;
     void subscribe(User user, ChatObserver observer, ChatLog chatLog) throws RemoteException;
